+++ src/python/pants/backend/go/goals/tailor.py
